requirements.txt
cs50
python-dotenv
Flask
Flask-Session
spotipy
